Move the hstore converter tests into their own util test class
