+++ src/test/java/io/confluent/connect/jdbc/dialect/PostgreSqlDatabaseDialectTest.java
@@ -22,7 +22,6 @@ import io.confluent.connect.jdbc.sink.metadata.SinkRecordField;
 import io.confluent.connect.jdbc.source.ColumnMapping;
 import io.confluent.connect.jdbc.source.JdbcSourceConnectorConfig;
 import io.confluent.connect.jdbc.util.ColumnDefinition;
-import io.confluent.connect.jdbc.util.HstoreConverter;
 import io.confluent.connect.jdbc.util.ColumnId;
 import io.confluent.connect.jdbc.util.QuoteMethod;
 import io.confluent.connect.jdbc.util.TableDefinition;
@@ -1257,98 +1256,6 @@ public class PostgreSqlDatabaseDialectTest extends BaseDialectTest<PostgreSqlDat
         new PostgreSqlDatabaseDialect(sinkConfigWithUrl("jdbc:postgresql://something"));
     assertThrows(ConnectException.class,
         () -> disabled.bindField(mock(PreparedStatement.class), 1, schema, value));
-  }
-
-  /**
-   * The hstore text form the bind produces. Everything is quoted, so a delimiter inside a key or
-   * value is inert and the string {@code NULL} stays distinct from a NULL value.
-   */
-  @Test
-  public void shouldSerializeMapsToHstoreText() {
-    Map<String, String> map = new LinkedHashMap<>();
-    map.put("env", "prod");
-    map.put("absent", null);
-    map.put("literal", "NULL");
-    map.put("a=>b", "c,d");
-    map.put("say \"hi\"", "back\\slash");
-
-    assertEquals("\"env\"=>\"prod\",\"absent\"=>NULL,\"literal\"=>\"NULL\","
-            + "\"a=>b\"=>\"c,d\",\"say \\\"hi\\\"\"=>\"back\\\\slash\"",
-        HstoreConverter.connectMapToHstore(map));
-
-    assertEquals("", HstoreConverter.connectMapToHstore(Collections.emptyMap()));
-    assertNull(HstoreConverter.connectMapToHstore(null));
-  }
-
-  /**
-   * The text form the driver hands back for an hstore it could not resolve. Quoting makes the
-   * delimiters inert, and a bare NULL is a null value while a quoted one is the string.
-   */
-  @Test
-  public void shouldParseHstoreText() {
-    Map<String, String> expected = new LinkedHashMap<>();
-    expected.put("env", "prod");
-    expected.put("absent", null);
-    expected.put("literal", "NULL");
-    expected.put("a=>b", "c,d");
-    expected.put("say \"hi\"", "back\\slash");
-
-    // PostgreSQL renders pairs separated by ", " and escapes quotes and backslashes.
-    assertEquals(expected, HstoreConverter.hstoreToConnectMap(
-        "\"env\"=>\"prod\", \"absent\"=>NULL, \"literal\"=>\"NULL\", "
-            + "\"a=>b\"=>\"c,d\", \"say \\\"hi\\\"\"=>\"back\\\\slash\""));
-
-    assertEquals(Collections.emptyMap(), HstoreConverter.hstoreToConnectMap(""));
-    assertEquals(Collections.emptyMap(), HstoreConverter.hstoreToConnectMap("   "));
-    assertNull(HstoreConverter.hstoreToConnectMap(null));
-  }
-
-  /** Whatever the serializer writes, the parser must read back identically. */
-  @Test
-  public void shouldRoundTripHstoreTextThroughBothDirections() {
-    Map<String, String> original = new LinkedHashMap<>();
-    original.put("env", "prod");
-    original.put("absent", null);
-    original.put("literal", "NULL");
-    original.put("a=>b", "c,d");
-    original.put("say \"hi\"", "back\\slash");
-    original.put("key 2", " ##123 78");
-    original.put("empty", "");
-
-    assertEquals(original, HstoreConverter.hstoreToConnectMap(
-        HstoreConverter.connectMapToHstore(original)));
-    assertEquals(Collections.emptyMap(), HstoreConverter.hstoreToConnectMap(
-        HstoreConverter.connectMapToHstore(Collections.emptyMap())));
-  }
-
-  @Test
-  public void shouldRejectMalformedHstoreText() {
-    for (String malformed : new String[]{
-        "not hstore at all",          // no => at all
-        "\"k\"",                      // key with no value
-        "\"k\"=>",                    // separator with no value
-        "\"k\"=>\"unterminated",      // unclosed quote
-        "\"a\"=>\"1\" \"b\"=>\"2\"",     // missing comma between pairs
-        "a=>1",                       // unquoted: hstore_out always quotes, so this is not ours
-        "\"a\"=>1",                   // unquoted value, likewise
-        "\"a\"=>null"                 // lowercase null is the string, never the NULL literal
-    }) {
-      assertThrows("should reject: " + malformed, DataException.class,
-          () -> HstoreConverter.hstoreToConnectMap(malformed));
-    }
-  }
-
-  @Test
-  public void shouldRejectValuesThatAreNotStringMaps() {
-    assertThrows(DataException.class, () -> HstoreConverter.connectMapToHstore("not a map"));
-    assertThrows(DataException.class,
-        () -> HstoreConverter.connectMapToHstore(Collections.singletonMap(null, "v")));
-    // A non-String key or value fails rather than being coerced through toString, so a schema and
-    // value that disagree surface here instead of silently reaching the database.
-    assertThrows(DataException.class,
-        () -> HstoreConverter.connectMapToHstore(Collections.singletonMap("k", 1)));
-    assertThrows(DataException.class,
-        () -> HstoreConverter.connectMapToHstore(Collections.singletonMap(1, "v")));
   }
 
   /**
